moved event to seperate class
